move nats stuff into `nats/nats.go`
