+++ app/router.go
@@ -4,12 +4,9 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"log/slog"
-	"net"
 	"net/http"
 	"northstar/config"
-	"os"
-	"strconv"
+	"northstar/nats"
 	"sync"
 	"time"
 
@@ -20,11 +17,8 @@ import (
 	"northstar/app/features/reverse"
 	"northstar/app/features/sortable"
 
-	"github.com/delaneyj/toolbelt"
-	"github.com/delaneyj/toolbelt/embeddednats"
 	"github.com/go-chi/chi/v5"
 	"github.com/gorilla/sessions"
-	natsserver "github.com/nats-io/nats-server/v2/server"
 	"github.com/starfederation/datastar-go/datastar"
 )
 
@@ -53,24 +47,10 @@ func SetupRoutes(ctx context.Context, router chi.Router) (err error) {
 		})
 	}
 
-	natsPort, err := getFreeNatsPort()
+	ns, err := nats.SetupNATS(ctx)
 	if err != nil {
-		return fmt.Errorf("error obtaining NATS port: %w", err)
+		return fmt.Errorf("error setting up NATS: %w", err)
 	}
-
-	ns, err := embeddednats.New(ctx, embeddednats.WithNATSServerOptions(&natsserver.Options{
-		JetStream: true,
-		NoSigs:    true,
-		Port:      natsPort,
-		StoreDir:  "data/nats",
-	}))
-
-	if err != nil {
-		return fmt.Errorf("error creating embedded nats server: %w", err)
-	}
-
-	ns.WaitForServer()
-	slog.Info("NATS started", "port", natsPort)
 
 	sessionStore := sessions.NewCookieStore([]byte("session-secret"))
 	sessionStore.MaxAge(int(24 * time.Hour / time.Second))
@@ -87,32 +67,4 @@ func SetupRoutes(ctx context.Context, router chi.Router) (err error) {
 	}
 
 	return nil
-}
-
-func getFreeNatsPort() (int, error) {
-	if p, ok := os.LookupEnv("NATS_PORT"); ok {
-		natsPort, err := strconv.Atoi(p)
-		if err != nil {
-			return 0, fmt.Errorf("error parsing NATS_PORT: %w", err)
-		}
-		if isPortFree(natsPort) {
-			return natsPort, nil
-		}
-	}
-	return toolbelt.FreePort()
-}
-
-func isPortFree(port int) bool {
-	address := fmt.Sprintf(":%d", port)
-
-	ln, err := net.Listen("tcp", address)
-	if err != nil {
-		return false
-	}
-
-	if err := ln.Close(); err != nil {
-		return false
-	}
-
-	return true
 }
